Try fixing faucet (untested)
With this, the faucetting is moved here, using a zkapp command precondition
to ensure that the recipient doesn't already have the funds.

Of course they can misuse the faucet by constantly emptying their account,
but this is no different than the abuse they can do by making new accounts.

The faucet is temporary anyway.